requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
compressai==1.1.5
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
fonttools==4.40.0
fsspec==2023.6.0
huggingface-hub==0.15.1
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
lit==16.0.5.post0
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
Pillow==9.5.0
ptflops==0.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-msssim==1.0.0
PyYAML==6.0
requests==2.31.0
safetensors==0.3.1
scipy==1.10.1
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.3
urllib3==2.0.3
zipp==3.15.0
